LayoutTests/storage/indexeddb/intversion-omit-parameter-expected.txt

Test that initial version after a successful open of a non-existent db is 1

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-omit-parameter.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)

openSuccess():
db = event.target.result
Test line from IDBFactory.open: If no version is specified and no database exists, set database version to 1.
FAIL db.version should be 1 (of type number). Was  (of type string).
PASS successfullyParsed is true

TEST COMPLETE
